pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "momentum"
version = "2.0.0"
description = "A light-weight, local-first task tracking application"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    {name = "mek0124", email = "mek0124@proton.me"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Scheduling",
    "Topic :: Utilities",
]
dependencies = [
    "pyside6>=6.9.1",
    "qfluentwidgets>=1.10.5",
    "sqlalchemy>=2.0.45",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "markdown>=3.10.1",
    "coverage==7.13.3",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "pytest-qt==4.5.0",
    # API dependencies
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.1",
    "stripe>=7.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.8.0",
    "flake8>=7.1.1",
    "pytest>=8.3.2",
    "ruff>=0.7.2",
]

[project.urls]
Homepage = "https://github.com/mek0124/momentum"
Repository = "https://github.com/mek0124/momentum"
Issues = "https://github.com/mek0124/momentum/issues"

[tool.setuptools]
packages = [
    "app",
    "app.core",
    "app.database", 
    "app.models",
    "app.pages",
    "app.utils",
    "updater",
    "updater.pages",
    "updater.utils",
]

[tool.setuptools.package-dir]
app = "app"
"app.core" = "app/core"
"app.database" = "app/database"
"app.models" = "app/models"
"app.pages" = "app/pages"
"app.utils" = "app/utils"
updater = "updater"
"updater.pages" = "updater/pages"
"updater.utils" = "updater/utils"

[tool.setuptools.package-data]
app = ["assets/*.png"]
updater = ["pages/*.py", "utils/*.py"]

[project.scripts]
momentum = "main:run_main"

[project.entry-points."gui_scripts"]
momentum-gui = "main:run_main"